Encapsulate NodeState, Misbehaving; Inject chainActive as dependency
