+++ divi/src/main.h
@@ -106,8 +106,6 @@ CBlockIndex* InsertBlockIndex(uint256 hash);
 bool AbortNode(const std::string& msg, const std::string& userMessage);
 /** Get statistics from node state */
 bool GetNodeStateStats(int nodeid, CNodeStateStats& stats);
-/** Increase a node's misbehavior score. */
-void Misbehaving(int nodeid, int howmuch);
 /** Flush all state, indexes and buffers to disk. */
 void FlushStateToDisk();
